librbd: move all image operation state machines to new namespace
Prior to creating multiple new state machines for image operations,
move all existing op requests to a new namespace and folder.
