+++ src/librbd/ImageCtx.h
@@ -43,7 +43,10 @@ namespace librbd {
   class CopyupRequest;
   class ImageWatcher;
   class Journal;
+
+  namespace operation {
   class ResizeRequest;
+  }
 
   struct ImageCtx {
     CephContext *cct;
@@ -132,7 +135,7 @@ namespace librbd {
 
     atomic_t async_request_seq;
 
-    xlist<ResizeRequest*> resize_reqs;
+    xlist<operation::ResizeRequest*> resize_reqs;
 
     AioImageRequestWQ *aio_work_queue;
     ContextWQ *op_work_queue;
