+++ src/librbd/operation/FlattenRequest.h
@@ -1,7 +1,7 @@
 // -*- mode:C++; tab-width:8; c-basic-offset:2; indent-tabs-mode:t -*-
 // vim: ts=8 sw=2 smarttab
-#ifndef CEPH_LIBRBD_FLATTEN_REQUEST_H
-#define CEPH_LIBRBD_FLATTEN_REQUEST_H
+#ifndef CEPH_LIBRBD_OPERATION_FLATTEN_REQUEST_H
+#define CEPH_LIBRBD_OPERATION_FLATTEN_REQUEST_H
 
 #include "librbd/AsyncRequest.h"
 #include "librbd/parent_types.h"
@@ -11,6 +11,8 @@ namespace librbd {
 
 class ImageCtx;
 class ProgressContext;
+
+namespace operation {
 
 class FlattenRequest : public AsyncRequest
 {
@@ -76,6 +78,7 @@ private:
   bool send_update_children();
 };
 
+} // namespace operation
 } // namespace librbd
 
-#endif // CEPH_LIBRBD_FLATTEN_REQUEST_H
+#endif // CEPH_LIBRBD_OPERATION_FLATTEN_REQUEST_H
