+++ src/librbd/operation/TrimRequest.h
@@ -1,7 +1,7 @@
 // -*- mode:C++; tab-width:8; c-basic-offset:2; indent-tabs-mode:t -*-
 // vim: ts=8 sw=2 smarttab
-#ifndef CEPH_LIBRBD_TRIM_REQUEST_H
-#define CEPH_LIBRBD_TRIM_REQUEST_H
+#ifndef CEPH_LIBRBD_OPERATION_TRIM_REQUEST_H
+#define CEPH_LIBRBD_OPERATION_TRIM_REQUEST_H
 
 #include "librbd/AsyncRequest.h"
 
@@ -10,6 +10,8 @@ namespace librbd
 
 class ImageCtx;
 class ProgressContext;
+
+namespace operation {
 
 class TrimRequest : public AsyncRequest
 {
@@ -85,6 +87,7 @@ private:
   void finish(int r);
 };
 
+} // namespace operation
 } // namespace librbd
 
-#endif // CEPH_LIBRBD_TRIM_REQUEST_H
+#endif // CEPH_LIBRBD_OPERATION_TRIM_REQUEST_H
